GEN_AI_Tasks/Task 1/code/requirements.txt
feedparser
beautifulsoup4
google-generativeai
Pillow
moviepy
